CoinGun Implement & Dungeon Map Add

+++ MyEngine/MyEngine/BossMonster.cpp
@@ -74,7 +74,7 @@ void BossMonster::OnCollisionEnter(Collider* _other)
 
 		float curHP = GetCurHP();
 
-		int att = minAtt + (rand() % (int)(maxAtt - minAtt));
+		int att = (int)(minAtt + (rand() % (int)(maxAtt - minAtt)));
 
 		float damage = (att + playerAtt);
 		curHP -= damage;
